pyproject.toml
[tool.poetry]
name = "thechannelcompany"
version = "0.1.0"
description = "Daniel's programming tests"
authors = ["Daniel <daniel@danielpgleason.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.7"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.2"
markdown = "^3.6"
django-filter = "^24.2"
pillow = "^10.4.0"

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true

[tool.pylint.'MESSAGES CONTROL']
disable = [
    "import-error",  # E0401
    "missing-module-docstring",  # C0114
    "missing-class-docstring",   # C0115
    "missing-function-docstring"  # C0116
]

[tool.pylint]
max-line-length = 200

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"